# portal/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Compliance Engine API
  version: 1.0.0
  description: |
    API for the Compliance Engine platform — ASV scanning, WAF management,
    Wazuh SIEM integration, and compliance assessment (SOC 2, PCI DSS, PCI SSS).

    ## Authentication

    All API requests require an API key passed in the `X-API-Key` header.
    Keys are scoped — see the `scopes` object on key creation for available
    permissions.

    ## Scopes

    | Scope | Grants |
    |-------|--------|
    | `read:scans` | Read scan results and status |
    | `write:scans` | Start and stop scans |
    | `read:waf` | Read WAF rules and traffic stats |
    | `manage:waf` | Create, update, delete WAF rules |
    | `read:siem` | Read SIEM alerts and agents |
    | `write:siem` | Resolve alerts, trigger active response |
    | `read:compliance` | Read compliance frameworks and assessments |
    | `admin` | Full access to all resources |

  contact:
    name: Compliance Engine Support
    email: support@compliance-engine.example
  license:
    name: Proprietary
servers:
  - url: https://api.compliance-engine.example/v1
    description: Production
  - url: https://sandbox.compliance-engine.example/v1
    description: Sandbox
security:
  - ApiKeyAuth: []
...
paths:
...
  /scans:
    post:
      tags: [Scanners]
      summary: Start a scan
      operationId: createScan
      security:
        - ApiKeyAuth: [write:scans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanRequest'
      responses:
        '202':
          description: Scan accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Scanners]
      summary: List scans
      operationId: listScans
      security:
        - ApiKeyAuth: [read:scans]
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Scan list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scan'
  /scans/{scanId}:
    parameters:
      - $ref: '#/components/parameters/scanId'
    get:
      tags: [Scanners]
      summary: Get scan status and results
      operationId: getScan
      security:
        - ApiKeyAuth: [read:scans]
      responses:
        '200':
          description: Scan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Scanners]
      summary: Stop a running scan
      operationId: stopScan
      security:
        - ApiKeyAuth: [write:scans]
      responses:
        '204':
          description: Scan stopped
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    scanId:
      name: scanId
      in: path
      required: true
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    ScanType:
      type: string
      enum: [ASV, VULNERABILITY, COMPLIANCE]
    ScanStatus:
      type: string
      enum: [PENDING, RUNNING, COMPLETED, FAILED]
    Severity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    ScanRequest:
      type: object
      required: [target, type]
      properties:
        target:
          type: string
          example: https://example.com
        type:
          $ref: '#/components/schemas/ScanType'
        options:
          type: object
          additionalProperties: true
    Vulnerability:
      type: object
      properties:
        id:
          type: string
        severity:
          $ref: '#/components/schemas/Severity'
        description:
          type: string
        recommendation:
          type: string
    ScanResults:
      type: object
      properties:
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/Vulnerability'
        score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        summary:
          type: string
    Scan:
      type: object
      properties:
        id:
          type: string
        target:
          type: string
        type:
          $ref: '#/components/schemas/ScanType'
        status:
          $ref: '#/components/schemas/ScanStatus'
        results:
          $ref: '#/components/schemas/ScanResults'
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
...