# portal/src/server/scanners/tempest.ts
export interface ScanRequest {
  target: string;
  type: "ASV" | "VULNERABILITY" | "COMPLIANCE";
  options?: Record<string, unknown>;
}

export interface ScanResult {
  id: string;
  target: string;
  type: string;
  status: "PENDING" | "RUNNING" | "COMPLETED" | "FAILED";
  results?: {
    vulnerabilities: Array<{
      id: string;
      severity: string;
      description: string;
      recommendation: string;
    }>;
    score: number;
    summary: string;
  };
  startedAt: string;
  completedAt?: string;
}

export class T3MP3STClient {
  private apiUrl: string;
  private apiKey?: string;

  constructor(apiUrl: string, apiKey?: string) {
    this.apiUrl = apiUrl;
    this.apiKey = apiKey;
  }

  async startScan(request: ScanRequest): Promise<ScanResult> {
    // TODO: Implement T3MP3ST API integration
    // POST to ${this.apiUrl}/api/scan
    throw new Error("Not implemented - integrate with T3MP3ST");
  }

  async getScanStatus(scanId: string): Promise<ScanResult> {
    // TODO: Implement T3MP3ST API integration
    // GET to ${this.apiUrl}/api/scan/${scanId}
    throw new Error("Not implemented - integrate with T3MP3ST");
  }

  async stopScan(scanId: string): Promise<void> {
    // TODO: Implement T3MP3ST API integration
    // DELETE to ${this.apiUrl}/api/scan/${scanId}
    throw new Error("Not implemented - integrate with T3MP3ST");
  }
}